- Start of refactoring of reusable functions in RoomSelection and BookingReview
- Added useDatechange.ts and dataUtils.ts for reuse
